Add Timber for logs, remove and rename some duplicate class and transtalte spanish reference to english. close #7 and close #4
In App.kt was created another class ReleaseTree for reuse logs in release version,
in case you consider (@eleazan @juancoob @jujoru.)
Or if, on the contrary, we remove this possibility.
