+++ app/build.gradle
@@ -71,6 +71,9 @@ dependencies {
     androidTestImplementation 'androidx.test.ext:junit:1.1.2'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
 
+    // Gson
+    implementation 'com.google.code.gson:gson:2.8.6'
+
     // Retrofit
     implementation 'com.squareup.okhttp3:logging-interceptor:4.9.0'
     implementation 'com.squareup.retrofit2:retrofit:2.9.0'
@@ -80,4 +83,7 @@ dependencies {
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:1.4.2"
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:1.4.2"
     implementation "androidx.lifecycle:lifecycle-runtime-ktx:2.3.0"
+
+    // Timber
+    implementation 'com.jakewharton.timber:timber:4.7.1'
 }
