+++ app/src/main/java/com/architectcoders/arquitectomarvel/model/characters/Comics.kt
@@ -1,15 +1,15 @@
-package com.architectcoders.arquitectomarvel.model.personajes
+package com.architectcoders.arquitectomarvel.model.characters
 
 
 import com.google.gson.annotations.SerializedName
 
-data class Series(
+data class Comics(
     @SerializedName("available")
     val available: Int?,
     @SerializedName("collectionURI")
     val collectionURI: String?,
     @SerializedName("items")
-    val items: List<ItemXX>?,
+    val items: List<Item>?,
     @SerializedName("returned")
     val returned: Int?
 )
