+++ app/src/main/java/com/architectcoders/arquitectomarvel/model/characters/Events.kt
@@ -1,15 +1,15 @@
-package com.architectcoders.arquitectomarvel.model.personajes
+package com.architectcoders.arquitectomarvel.model.characters
 
 
 import com.google.gson.annotations.SerializedName
 
-data class Stories(
+data class Events(
     @SerializedName("available")
     val available: Int?,
     @SerializedName("collectionURI")
     val collectionURI: String?,
     @SerializedName("items")
-    val items: List<ItemXXX>?,
+    val items: List<Item>?,
     @SerializedName("returned")
     val returned: Int?
 )
