+++ app/src/main/java/com/architectcoders/arquitectomarvel/model/characters/Stories.kt
@@ -1,15 +1,15 @@
-package com.architectcoders.arquitectomarvel.model.personajes
+package com.architectcoders.arquitectomarvel.model.characters
 
 
 import com.google.gson.annotations.SerializedName
 
-data class Events(
+data class Stories(
     @SerializedName("available")
     val available: Int?,
     @SerializedName("collectionURI")
     val collectionURI: String?,
     @SerializedName("items")
-    val items: List<ItemX>?,
+    val itemsStories: List<ItemStories>?,
     @SerializedName("returned")
     val returned: Int?
 )
