+++ app/src/main/java/com/architectcoders/arquitectomarvel/model/characters/Thumbnail.kt
@@ -1,4 +1,4 @@
-package com.architectcoders.arquitectomarvel.model.personajes
+package com.architectcoders.arquitectomarvel.model.characters
 
 
 import com.google.gson.annotations.SerializedName
