--- app/src/main/java/com/architectcoders/arquitectomarvel/model/personajes/Item.kt
@@ -1,12 +1,0 @@
-package com.architectcoders.arquitectomarvel.model.personajes
-
-
-import com.google.gson.annotations.SerializedName
-
-data class Item(
-    @SerializedName("name")
-    val name: String?,
-    @SerializedName("resourceURI")
-
-    val resourceURI: String?
-)
